requirements-dev.txt
alabaster==0.7.12
amqp==2.6.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.4.1
astroid==2.8.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.1.7
billiard==3.6.4.0
black==21.9b0
blinker==1.4
Brlapi==0.7.0
celery==4.4.6
certifi==2019.11.28
cffi==1.14.6
cfgv==3.3.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.0.1
colorama==0.4.3
command-not-found==0.3
coverage==5.5
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.8
django-allauth==0.45.0
django-celery-beat==2.2.0
django-coverage-plugin==2.0.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-filter==21.1
django-model-utils==4.1.1
django-redis==5.0.0
django-restframework==0.0.1
django-stubs==1.8.0
django-stubs-ext==0.3.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
docutils==0.17.1
duplicity==0.8.12.0
entrypoints==0.3
factory-boy==3.2.0
Faker==8.14.0
fasteners==0.14.1
filelock==3.0.12
flake8==3.9.2
flake8-isort==4.0.0
flower==0.9.7
future==0.18.2
GDAL==3.0.4
h11==0.12.0
hiredis==2.0.0
httplib2==0.14.0
httptools==0.2.0
humanize==3.11.0
identify==2.2.15
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.28.0
isort==5.9.3
jedi==0.18.0
Jinja2==3.0.1
keyring==18.0.1
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
libvirt-python==6.1.0
livereload==2.6.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
mypy==0.910
mypy-extensions==0.4.3
netifaces==0.10.4
nodeenv==1.6.0
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==21.0
paramiko==2.6.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.3.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.20
protobuf==3.6.1
psycopg2==2.9.1
py==1.10.0
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==1.9.73
pyflakes==2.3.1
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.11.1
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==6.2.4
pytest-django==4.4.0
pytest-sugar==0.9.4
python-apt==2.0.0+ubuntu0.20.4.4
python-crontab==2.5.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.5
python-dotenv==0.19.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
regex==2021.8.28
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
static3==0.7.0
systemd-python==234
termcolor==1.1.0
testfixtures==6.18.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.14.0
uvloop==0.16.0
vboxapi==1.0
vine==1.3.0
virtualenv==20.0.17
wadllib==1.3.3
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
